Refactored auth code.
