+++ lib/auth/sign_in.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const Settings = require('./settings');
-const AuthUtils = require('./auth_utils');
+const Settings = require('../settings');
+const AuthUtils = require('./utils');
 
 const clientId = '468459411007078400';
 const appId = 'AKfycby3tLs26ufCxJX6LaUaGNUfi4RsGAqlIOMQX8_reYJu2v59Voj9';
