+++ lib/auth/sign_in_token.js
@@ -1,8 +1,8 @@
 'use strict';
 
-const Settings = require('./settings');
-const Utils = require('./utils');
-const AuthUtils = require('./auth_utils');
+const Settings = require('../settings');
+const Utils = require('../utils');
+const AuthUtils = require('./utils');
 
 $(() => {
   const accessToken = Utils.urlParameter('accessToken');
