+++ lib/auth/utils.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Settings = require('./settings');
+const Settings = require('../settings');
 
 const DISCORD_API_URL = 'https://discordapp.com/api';
